T5R2CorrectingFilterInterpreter: Fixes some T5R2 input.
On T5R2 pads, we occasionally see input events that specify a non-zero
touch count, followed by no further events. We always see two of these
inputs with identical touch_cnt and finger_cnt in a row. It looks like
a bug in the firmware for these devices. Here is some data we've
collected for a 3 finger tap:

legend: touch_cnt [finger_ids]

3 [251 252]
3 [251 252]
3 [251 252]
1 []
1 []

and:

3 [257 258]
1 []
1 []

This change looks for two packets w/ finger_cnt == 0 and identical
non-zero touch_cnt, and in such cases sets the second hardware state's
touch_cnt to 0.

BUG=chromium-os:25973
TEST=unittest; tested on device

Change-Id: I89aa0c6b1de0386be8ec554adc86afea6b3adef9