ImmediateInterpreter: Allow finger pairs to be slightly further apart.
We used to require two fingers to be 40mm apart or closer to perform a
two finger gesture (like right click). However, during a test with Ian
Fette, who has big hands, we found him having a lot of trouble right
clicking. This change increases the distance required for a right
click or any other two finger gesture.

We also take part of the log from Ian's test and turn it into a
unittest.

BUG=chromium-os:26945
TEST=unittest

Change-Id: I66d523c802ebd873139b260ed6b055e4b3bc2c4a