ImmediateInterpreter: When going to idle state, clear tap record.
We should clear the tap record when going to idle state, but there was
a case where we didn't if the physical button was down. This could
lead to a situation where a tap starts, then the physical button
clicks (sending state to idle, but not clearing tap record). After the
physical button goes back up, we see we are in idle state and b/c of
the old tap record, we see a tap has begun and go to the next state
(tap began). Then when the fingers let go a tap occurs! End result: a
single 2f click yields a 2f click and then a 2f tap click.

This CL moves clearing the tap record to the case where it's first
used in the idle state, thus making sure it's cleared no matter how we
got to the idle state.

BUG=chromium-os:27485
TEST=unittest

Change-Id: I0cb002a7489108bc5d75fcbad6739d71f17afb51