ImmediateInterpreter: Use an ellipse to know if fingers are gesturing together
Historically we've used a circle to see if fingers are gesturing. That
is, we used a simple test to see how far apart two fingers were, and
if they were under that threshold, we considered them for two finger
gesturing. We've had a number of issues filed where people were
accidentally getting right clicks and their two fingers had a lot of
vertical space. Thus, we switch to an ellipse. We keep the same
horizontal distance as before (50mm), but reduce the vertical to 30mm.

BUG=chromium-os:29401
TEST=unittest; replayed kan's and mdhayter's logs to see issue gone

Change-Id: I91a314e2789e6c9829fa9fea819051929c0d6b97
Reviewed-on: https://gerrit.chromium.org/gerrit/20280
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/20341