LookaheadFilterInterpreter and others: drumroll-separated fingers can't tap.
One problem we've noticed for fast swipes is that, if drumroll
detection kicks in and the swipe is separated into separate finger
IDs, some of them may look like taps.

A previous CL helped address this problem by using variable delays to
avoid unnecessary separations, but they can still occur if we only get
two frames of data (which would be a rather fast swipe).

This CL introduces a new finger flag, similar to the two existing WARP
flags, which marks a finger as unable to cause a tap gesture. When
separating a finger, we set this on both sides of the separation. To
help ensure that we didn't already execute the earlier half for fast
swipes, we now use variable delay on a new finger that touches down.

This CL also updates ImmediateInterpreter to honor the new flag.

BUG=chromium-os:29576
TEST=unittest; tested fast swipes on device and saw no taps occur.

Change-Id: I83e25e66589608f5640fa45f69e2842eadcf69e8
Reviewed-on: https://gerrit.chromium.org/gerrit/20388
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/20886