LookaheadFilterInterpreter: Bias toward future when combining gestures.
Previous to this CL, the behavior when combining gestures was to just
use the older one when it came to scroll vs pointing. However, that
led to a bug where a scroll looked a bit like pointing at first, and
at the time we combined gestures we had an older move gesture and a
newer scroll gesture. In this case, we should go with the more recent
gesture.

BUG=chromium-os:27979
TEST=unittest; replayed problematic log

Change-Id: I00ad412e572ccee6caf21984d390495250d99a5b
Reviewed-on: https://gerrit.chromium.org/gerrit/20769
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/20973
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>