ImmediateInterpreter: Handle right taps that exhibit some "light" merging
We noticed on some hardware that when two fingers join or leave,
sometimes there is a frame of data that has only one contact, toward
the center of where the two are in other frames, and this contact has
very light pressure. This CL uses that very light pressure as a guage
that these frames shouldn't count toward moving a lot, since such
movement would cause a tap to look like it's not a tap.

BUG=chromium-os:30097
TEST=unittest, which is based on logs that failed, but now pass

Change-Id: I9d0e9fca4151f14b69fce713ad332183879ee635
Reviewed-on: https://gerrit.chromium.org/gerrit/21397
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/21552
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>