ImmediateInterpreter: Preserve Palm state when fingers change.
Historically we reevaluated whether a finger was a palm whenever any
finger arrived or left the pad. With this CL, we will not keep the
palm classification around for the lifetime of the contact. When
paried with a future CL, this will help palm detection considerably.

BUG=chromium-os:30615
TEST=unittest; problematic log was replayed and showed better
behavior.

Change-Id: I8f0277605094b0bd255b0bd5b3e70b434fd0a87e
Reviewed-on: https://gerrit.chromium.org/gerrit/22304
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/23280