ImmediateInterpreter: contacts in palm zone must move to cause scroll.
This is an extension of the behavior that requires contacts in the
bottom dampened zone to move to cause scroll (to help distinguish
between them and resting fingers/thumb). With this CL we make a
similar requirement of possible palms. I've received multiple logs
lately that show palms accidentally contributing to scroll.

BUG=chromium-os:30615
TEST=unittest; replayed problematic to see movement, not scroll; tried
on device.

Change-Id: I82fb409a1b2af1e6324daa5035d0ebb732ab075c
Reviewed-on: https://gerrit.chromium.org/gerrit/22306
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/23282
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>