SemiMtCorrectingFilterInterpreter: Suppress one-finger jump
The patch tries to address the big cursor jumps caused by drumroll-like
gestures with one-finger input samples only. For example, a previously
stationary (or very slowly moving, i.e. motion < move_threshold) single
finger that suddenly appears to jump by a large distance (> jump_threshold)
looks suspiciously like drum roll. When we detect this, report its old
position, but still save the amount that it moved. If the next sample shows
that it has not continued to move at a reasonable speed (motion < half of
the jump distance), then we assume that the jump was caused by drumroll, and
report it as a new finger at its new position with a new tracking id. On the
other hand, if the finger keeps moving far away or close to the old position,
just report the new position with the same tracking id.

BUG=chromium-os:31423
TEST=unittest passed

Change-Id: I856300320ac00663ae679d755f84596109a60aaa
Reviewed-on: https://gerrit.chromium.org/gerrit/24874
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>