ImmediateInterpreter: Introduce "Fling Buffer Depth" property
The Fling Gesture velocity is computed from the last few scroll events.
These scroll events are stored in a buffer.  The default size (3) of the
buffer was being to determine the max number of samples to use when
computing the velocity.

For systems that don't use interpolation, there is a 1:1 mapping between
hardware states and scroll events in the buffer.
However, for systems that use 2x interpolation, such as synaptics T5R2
and SemiMt, the last 3 scrolls corresponded to the effect of only 1.5
hardware states.

To get a more realistic fling, increase the buffer size to 6, and allow a
system that uses interpolation to set a flag (Fling Buffer Depth) to
override the default fling depth (3).

BUG=chromium-os:31960
TEST=unittests

Reviewed-on: https://gerrit.chromium.org/gerrit/25702
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
(cherry picked from commit 5cff9f7d8accfd3a19204631bcf3652db8f16e14)

Change-Id: I6967eb3ed20b80e5d8a2e91eccc56bed006a9706
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/25987
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>