PalmClassifyingFilterInterpreter: Allow finger move at palm envelope
There are some complains about the trackpad edge filter area is
too large and cursor movement at the trackpad edge is not working.

In this CL we add additional logic for classifying whether a finger
in the palm envelope is pointing:

1. The finger has never travelled out of the palm envelope.
(otherwise we fallback to the original palm classification logic
so that we constrain this CL's impact only to those fingers moving
within the palm envelope).

2.We add two threshold values
palm_pointing_min_dist_ (default 8.0)
palm_pointing_max_reverse_dist_ (default 0.3)

Ideally, we want a finger to be pointing only if it moves in one
direction significantly without zig-zag. But due to touch sensor's
inaccuracy, we make the rule to be that a finger has to move in
one direction significantly (> palm_pointing_min_dist_)
and with little move in the opposite direction
(< palm_pointing_max_reverse_dist_)

BUG=chrome-os-partner:14450
TEST=unittests pass. regression tests pass.
     1F movement at the trackpad edge works.
     No significant regression in palm causing unwated cursor
     movement.

Change-Id: I4a1b141700dc5c217ac1f62eae1019f0a2b96c89
Reviewed-on: https://gerrit.chromium.org/gerrit/34588
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Ready: Yufeng Shen <miletus@chromium.org>
Tested-by: Yufeng Shen <miletus@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/35170