ImmediateInterpreter: Prevent stationary fingers from flinging.
We do attempt to prevent lifting fingers from causing fling, but on
some platforms it's hard to detect lifting fingers, and the detection
is not perfect on any platform. Thus, we would often do small flings
when fingers were lifted. For a long time this wasn't a big issue, but
a recent change to the fling curve exacerbated the problem.

With this CL, we look for stationary fingers. In practice, it seems
that fingers that move under 10mm/sec for a little while before the
fling can be considered stationary.

BUG=chromium-os:35066
TEST=regression test: lumpy/scroll_without_fling

Change-Id: Ia3e5acb540cdd4e9fcf0ac51f7b3abd2d0e395e4
Reviewed-on: https://gerrit.chromium.org/gerrit/35223
Reviewed-by: Yufeng Shen <miletus@chromium.org>
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/35372
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>