ImmediateInterpreter: Respond to taps sooner when followed by right-tap
Historically, if we got this sequence of events:
- 1 finger tap
- 1 finger down
- second finger down (part of 2f tap)
- both release

We would wait for both to release before sending the initial tap. This
is unnecessary because there is no chance of getting into a drag
condition. With this fix, we send the tap as soon as we know that the
subsequent tap is not a left-tap, which could be drag.

BUG=chromium-os:35697
TEST=unittest; regression tests show no score change. A future
gestures change would have resulted in a regression test getting a
lower score, but this keeps the scores constant.

Change-Id: If8c2eaf15e8474ebd0d9ed0d4530b6780adc7b9e
Reviewed-on: https://gerrit.chromium.org/gerrit/36965
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>