Salsa: Update the try_touch_experiment UI
This changes the way the participants actually apply the treatments.
Previously, they had to copy paste each treatment and it became clear
that this was a huge pain.  Similarly, the UI once you had copied it
into Crosh was confusing.

This change replaces the individual treatment commands with one giant
command you copy for the entire experiment.  This command includes
information on *all* of the treatments and if you pass it into the new
try_touch_experiment program it will allow you to switch back and
forth between the treatments however you please with a simple curses
interface.  Beyond UI, it checks repeatedly now to make sure that
everything was applied correctly, hopefully getting rid of any errors in
how the treatments are viewed.

Hopefully this makes it easier for users to compare the experiments and
flip back and forth between them, etc.

try_touch_experiment will replace the command that is currently just
a function in Crosh.  Crosh will be patched so that the command now just
runs this instead.

BUG=chromium-os:38637
TEST=For the server-side changes, a dev server is running at
cmooney.mtv:8081 where you can see the differences.  Manual testing of
the new try_touch_experiment was done on a Link to confirm it works there.

Original-Change-Id: Ia0090629e8d988e8f517ff83ceac03a9bf808c19
Originally-Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Originally-Reviewed-on: https://gerrit.chromium.org/gerrit/42772
(cherry picked from commit 83aa6c2daa67332c573a18e3eaf5c091ebd7222b)

Change-Id: I26350be8f4c3e0a48a50d49e6f5b190dfd12b10a
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/44051
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>