activity_log.cc: Avoid crash
Sometimes, for reasons not known, when converting the internal data
structure of a log to JSON, we have a HardwareState that has a
finger_cnt but NULL fingers. This was leading to a crash.

We now check for this possibility and log an error.

Hopefully in the future we can find instances of this log and submit
fixes.

BUG=chromium:298288
TEST=No change in regression test scores; unitteests pass

Change-Id: I5b87e093ff733bfa67077fc514eceb7529663666
Reviewed-on: https://chromium-review.googlesource.com/171082
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Commit-Queue: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>