ActivityLog: Handle incoming HardwareState with too many fingers better.
Historically, when copying a HardwareState object to the log, we'd
copy all elements, then fix-up the fingers pointer. If there were more
fingers than expected coming in, we'd set the fingers pointer to
NULL. The problem is, if that happens, finger_cnt is likely > 0, which
will lead to a problem when it goes to read the log later: fingers is
NULL but finger_cnt is > 0.

To fix, when we set fingers to NULL, we now also set finger_cnt to 0.

BUG=chromium:298288
TEST=this:
- Put 6 fingers on Yoshi touchpad
- file feedback
- didn't crash X

Change-Id: Ibcd46fe7122052030146d72c17371b7c13020dec
Reviewed-on: https://chromium-review.googlesource.com/171083
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Commit-Queue: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>