ScalingFilterInterpreter: Special-case small mice moves
When using mice, it's frustrating when moving one device pixel moves
the cursor > 1 pixels on the screen. This CL avoids this problem by
looking for the special case of exactly 1 device pixel move, then
reducing the motion to just 1 pixel of motion.

We actually go a little under 1 pixel because it will be increased on
external displays later in the stack.

BUG=chromium:412864
TEST=logitech-t620/low_dpi_move regression test

Change-Id: Ia821be43b7ef20b6df69aee53034c75a001482ba
Reviewed-on: https://chromium-review.googlesource.com/219153
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Commit-Queue: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>