Prevent erroneous pinches when pinch locked out
The pinch_locked_ flag is used to indicate that a decision about whether
to pinch has been made is locked in for the currently gesturing fingers.
However, in UpdatePinchState we don't check whether the pinch is locked
in or locked out.  This would cause erroneous pinches with -nan scaling
in some edge cases.

This CL prevents those erroneous pinches.

BUG=b:72244799
TEST=use touchtests to check that this change fixes bad feedback
reports.

Change-Id: I63827fec75a86dac3d210db65144de0f28ea7280
Reviewed-on: https://chromium-review.googlesource.com/916624
Commit-Ready: Sean O'Brien <seobrien@chromium.org>
Tested-by: Sean O'Brien <seobrien@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>