Allow use of fake timestamps in gesture library
There are cases when touchpad timestamp deltas will not be reliable,
such as for bluetooth touchpads.  For example the timestamp delta may be
very small for several reports, then jump a lot to catch up.

In order to improve finger velocity tracking for these touchpads, we
ignore the timestamp delta reported by the device, and add a fake
timestamp delta based on the device's report rate.

BUG=b:117569252
TEST=Deploy to nocturne and test using bluetooth touchpad, with
CL 1274705

Change-Id: Ica2ff004553c997fd0334219767372f82d2ccac8
Reviewed-on: https://chromium-review.googlesource.com/1274765
Commit-Ready: Sean O'Brien <seobrien@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>