Increase timeout when fingers leave
Sometimes when scrolling or swiping with multiple fingers, one finger
will leave the edge of the touchpad before the others, causing the
gesture with fewer fingers to be detected. This CL increases the timeout
after an actively gesturing finger leaves before another gesture can be
detected.

It also now only suspends gesture detection when it is an actively
gesturing finger that leaves.  If a non-gesturing thumb or palm leaves
the touchpad, gesture detection is not delayed.

BUG=chromium:986859
TEST=manual testing, and testing against regression test suite

Change-Id: I6dbff4766575f99ffd6ca86b6919d980014e2a27
Reviewed-on: https://chromium-review.googlesource.com/1716139
Tested-by: Sean O'Brien <seobrien@chromium.org>
Commit-Ready: Sean O'Brien <seobrien@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Reviewed-by: Harry Cutts <hcutts@chromium.org>